requirement.txt
attrdict==2.0.1
certifi==2019.3.9
chardet==3.0.4
cytoolz==0.9.0.1
Django==2.2.2
django-extensions==2.1.7
django-shell-plus==1.1.7
djangorestframework==3.9.4
eth-abi==1.3.0
eth-account==0.3.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.3
eth-rlp==0.1.2
eth-typing==2.1.0
eth-utils==1.6.0
hexbytes==0.2.0
idna==2.8
lru-dict==1.1.6
parsimonious==0.8.1
pycryptodome==3.8.2
pytz==2019.1
requests==2.22.0
rlp==1.1.0
six==1.12.0
sqlparse==0.3.0
toolz==0.9.0
urllib3==1.25.3
web3==4.9.2
websockets==6.0